requirements.txt
blinker==1.9.0
click==8.3.0
cloudinary==1.41.0
colorama==0.4.6
dnspython==2.8.0
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mongoengine==0.29.1
mongomock==4.3.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.3
pytest==8.4.2
pytest-flask==1.3.0
python-dotenv==1.2.1
pytz==2025.2
sentinels==1.1.1
Werkzeug==3.1.3
